pyproject.toml
[tool.poetry]
name = "idaeho"
version = "0.1.0"
description = ""
authors = ["emmanueluwa <oemmanuel.uwadiae@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.121.0"}
sqlalchemy = "^2.0.45"
psycopg2-binary = "^2.9.11"
python-dotenv = "^1.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
